tendermint: reorg mod crypto
Move CryptoProvider definition out of the crypto module into
the provider sub-module. Rename default_provider to default.
